throw error if no compiler
