+++ examples/tailwind/paperclip.config.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "target": "react",
     "outDir": "lib",
     "importAssetsAsModules": true
   },
