+++ packages/paperclip-builder/src/errors.ts
@@ -1,0 +1,3 @@
+import { createErrorClass } from "./utils";
+
+export const TargetNotFoundError = createErrorClass("TargetNotFoundError");
