+++ packages/paperclip-cli-minimal/src/build.ts
@@ -1,5 +1,9 @@
 import * as resolve from "resolve";
-import { buildDirectory } from "paperclip-builder";
+import {
+  buildDirectory,
+  TargetNotFoundError,
+  DirectoryBuilder
+} from "paperclip-builder";
 import * as path from "path";
 import * as URL from "url";
 import * as fs from "fs";
@@ -25,14 +29,26 @@ export type BuildOptions = {
 export const build = async (options: BuildOptions) => {
   const config = loadConfig(options);
   const engine = createEngineDelegate({});
-  const builder = buildDirectory(
-    {
-      watch: options.watch,
-      cwd: options.cwd,
-      config
-    },
-    engine
-  );
+
+  let builder: DirectoryBuilder;
+
+  try {
+    builder = buildDirectory(
+      {
+        watch: options.watch,
+        cwd: options.cwd,
+        config
+      },
+      engine
+    );
+  } catch (error) {
+    if (error instanceof TargetNotFoundError) {
+      console.error(error.message);
+    } else {
+      console.error(error);
+    }
+    return;
+  }
 
   builder
     .onFile((outFilePath: string, content: string) => {
