+++ packages/paperclip-coverage/package.json
@@ -5,10 +5,9 @@
   "main": "lib/index.js",
   "types": "lib/index.ds.ts",
   "scripts": {
-    "build": "yarn build:definition && yarn build:paperclip && tsc",
-    "build:paperclip": "paperclip-minimal build --only=js,js.map,css --verbose",
-    "build:definition": "paperclip-minimal build --only=d.ts --output=src --verbose",
-    "build:watch": "concurrently \"tsc --watch\" \"yarn build:paperclip --watch\"  \"yarn build:definition --watch\""
+    "build": "yarn build:paperclip && tsc",
+    "build:paperclip": "paperclip-minimal build",
+    "build:watch": "yarn build:paperclip && concurrently \"tsc --watch\" \"yarn build:paperclip --watch\" "
   },
   "author": "",
   "license": "ISC",
