+++ packages/paperclip-diff/package.json
@@ -7,10 +7,9 @@
     "paperclip-diff": "./bin/paperclip-diff"
   },
   "scripts": {
-    "build": "yarn build:definition && yarn build:paperclip && tsc",
-    "build:definition": "paperclip-minimal build --only=d.ts --output=src",
+    "build": "yarn build:paperclip && tsc",
     "build:paperclip": "paperclip-minimal build",
-    "build:watch": "concurrently \"tsc --watch\" \"yarn build:definition --watch\" \"yarn build:paperclip --watch\""
+    "build:watch": "concurrently \"tsc --watch\" \"yarn build:paperclip --watch\""
   },
   "author": "",
   "license": "ISC",
