+++ packages/paperclip-website/paperclip.config.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-    "target": ["react"]
+    "target": "react"
   },
   "srcDir": "./src"
 }
